+++ src/get-config-base.js
@@ -4,6 +4,16 @@ module.exports = () => ({
   env: {
     browser: true,
   },
+  rules: {
+    'no-console': [
+      'error',
+      {
+        allow: [
+          'error',
+        ],
+      },
+    ],
+  },
   overrides: [
     {
       files: [